+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Improve logging during startup for better troubleshooting.
   - Refactor allowing calls to `clojure-lsp.main/run!` for manually passing args, useful for `lein-clojure-lsp` for example.
   - Internal: Move graalvm configuration to sqlite-jdbc.
+  - Recognize `deftest` as function definition form
   
 - API/CLI
   - Use clj-kondo custom lint for API as well, required for correct diagnostics API feature.